fix deletion of a testCase

+++ backend/controllers/testCase.js
@@ -66,31 +66,29 @@ testRouter.put('/:id', upload.fields([{ name: 'controlImage', maxCount: 1 }, { n
     if (request.user.admin) {
         try {
             const testToUpdate = {
-                id: request.params.id,
                 name: request.body.name,
                 type: request.body.type,
                 bacteriaSpecificImages: []
             }
             if (request.files) {
                 if (request.files.controlImage) {
-                    test.controlImage = { data: Buffer.from(request.files.controlImage[0].buffer).toString('base64'), contentType: request.files.controlImage[0].mimetype }
+                    testToUpdate.controlImage = { data: Buffer.from(request.files.controlImage[0].buffer).toString('base64'), contentType: request.files.controlImage[0].mimetype }
                 }
                 if (request.files.positiveResultImage) {
-                    test.positiveResultImage = { data: Buffer.from(request.files.positiveResultImage[0].buffer).toString('base64'), contentType: request.files.positiveResultImage[0].mimetype }
+                    testToUpdate.positiveResultImage = { data: Buffer.from(request.files.positiveResultImage[0].buffer).toString('base64'), contentType: request.files.positiveResultImage[0].mimetype }
                 }
                 if (request.files.negativeResultImage) {
-                    test.negativeResultImage = { data: Buffer.from(request.files.negativeResultImage[0].buffer).toString('base64'), contentType: request.files.negativeResultImage[0].mimetype }
+                    testToUpdate.negativeResultImage = { data: Buffer.from(request.files.negativeResultImage[0].buffer).toString('base64'), contentType: request.files.negativeResultImage[0].mimetype }
                 }
                 if (request.files.bacteriaSpecificImages) {
                     for (let i = 0; i < request.files.bacteriaSpecificImages.length; i++) {
                         const file = request.files.bacteriaSpecificImages[i]
                         const bacterium = await Bacterium.findOne({ name: file.originalname.substring(0, file.originalname.indexOf('.')) })
-                        test.bacteriaSpecificImages.push({ data: Buffer.from(file.buffer).toString('base64'), contentType: file.mimetype, bacterium })
+                        testToUpdate.bacteriaSpecificImages.push({ data: Buffer.from(file.buffer).toString('base64'), contentType: file.mimetype, bacterium })
                     }
                 }
             }
-            await Test.findOneAndUpdate(request.params.id, testToUpdate, { new: true, context: 'query' })
-            const updatetTest = await Test.findById(request.params.id).populate('Bacterium', { name: 1 })
+            const updatetTest = await Test.findByIdAndUpdate(request.params.id, testToUpdate, { new: true, runValidators: true, context: 'query' })
             return response.status(200).json(updatetTest)
         } catch (error) {
             return response.status(400).json({ error: error.message })
@@ -103,22 +101,31 @@ testRouter.put('/:id', upload.fields([{ name: 'controlImage', maxCount: 1 }, { n
 testRouter.delete('/:id', async (request, response) => {
     if (request.user.admin) {
         try {
-            const testToDelete = await Test.findById(request.params.id)
-            try {
-                const cases = await Case.find({ 'testGroups.test': testToDelete }).populate({
-                    path: 'testGroups.test',
-                    model: 'Test',
-                    populate: {
-                        path: 'bacteriaSpecificImages.bacterium',
-                        model: 'Bacterium'
-                    }
-                })
-                console.log('cases',cases)
-                if (cases > 0) {
-                    return response.status(400).json({ error: 'Testiä ei voida poistaa, koska se on jo käytössä tapauksessa' })
+            const testToDelete = await Test.findById(request.params.id).populate({
+                path: 'bacteriaSpecificImages.bacterium',
+                model: 'Bacterium'
+            })
+
+            const cases = await Case.find({ }).populate({
+                path: 'testGroups.test',
+                model: 'Test',
+                populate: {
+                    path: 'bacteriaSpecificImages.bacterium',
+                    model: 'Bacterium'
                 }
-            } catch (error){
-                return response.status(400).json({ error: error.message })
+            })
+            let testIsInUse = false
+            cases.forEach(element => {
+                for (let i = 0; i < element.testGroups.length; i++) {
+                    for (let j = 0; j < element.testGroups[i].length; j++) {
+                        if (element.testGroups[i][j].test.name === testToDelete.name) {
+                            testIsInUse = true
+                        }
+                    }
+                }
+            })
+            if (testIsInUse) {
+                return response.status(400).json({ error: 'Testi on käytössä ainakin yhdessä taupaksessa, eikä sitä voida poistaa' })
             }
 
             await Test.findByIdAndRemove(request.params.id)
